+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### v0.6.1
 
+* New is_path_newer command.
 * New get_last_modified_time command.
 
 ### v0.6.0 (2020-07-03)